Refactored shaderc such that its functionality can be used in runtime
